change to delete task center

+++ marketgo-server/mktgo-web/src/main/java/com/easy/marketgo/web/service/wecom/impl/WeComTaskCenterServiceImpl.java
@@ -427,7 +427,9 @@ public class WeComTaskCenterServiceImpl implements WeComTaskCenterService {
                         mediaUuidList.add(item.getLink().getMediaUuid());
                     }
                 });
-                weComMediaResourceRepository.deleteByUuids(mediaUuidList);
+                if (CollectionUtils.isNotEmpty(mediaUuidList)) {
+                    weComMediaResourceRepository.deleteByUuids(mediaUuidList);
+                }
             }
         }
